# internal/api/docs/swagger.yaml
definitions:
  company.BranchResponse:
    properties:
      address:
        type: string
      branch_id:
        type: string
      company_id:
        type: string
      created_at:
        type: string
      name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
...
  company.CreateBranchRequest:
    properties:
      address:
        type: string
      company_id:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  company.ListBranchesResponse:
    properties:
      branches:
        items:
          $ref: '#/definitions/company.BranchResponse'
        type: array
    type: object
...
  company.Message:
    properties:
      message:
        type: string
    type: object
  company.UpdateBranchRequest:
    properties:
      address:
        type: string
      branch_id:
        type: string
      company_id:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  entity.Error:
    properties:
      message:
        type: string
    type: object
...
  products.Category:
    properties:
      branch_id:
        description: Added branch_id
        type: string
      company_id:
        description: Company ID added
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
    type: object
  products.CategoryList:
    properties:
      categories:
        items:
          $ref: '#/definitions/products.Category'
        type: array
    type: object
...
  products.Error:
    properties:
      error:
        type: string
    type: object
...
  products.Message:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server
  title: API Gateway
  version: "1.0"
paths:
  /branches/{branch_id}:
    delete:
      consumes:
      - application/json
      description: Delete a branch by ID
      parameters:
      - description: Branch ID
        in: path
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/company.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Branch
      tags:
      - Branches
    get:
      consumes:
      - application/json
      description: Get branch details by ID
      parameters:
      - description: Branch ID
        in: path
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/company.BranchResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Branch
      tags:
      - Branches
    put:
      consumes:
      - application/json
      description: Update branch details
      parameters:
      - description: Branch ID
        in: path
        name: branch_id
        required: true
        type: string
      - description: Updated branch details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/company.UpdateBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/company.BranchResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Branch
      tags:
      - Branches
  /branches/create:
    post:
      consumes:
      - application/json
      description: Create a new branch
      parameters:
      - description: Branch details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/company.CreateBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/company.BranchResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create Branch
      tags:
      - Branches
  /branches/list:
    get:
      consumes:
      - application/json
      description: List all branches for a company
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/company.ListBranchesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Branches
      tags:
      - Branches
...
  /products/category:
    get:
      consumes:
      - application/json
      description: Retrieve a list of product categories
      parameters:
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      - description: Filter by category name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            $ref: '#/definitions/products.CategoryList'
        "400":
          description: Bad request due to invalid query parameters
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: List Product Categories
      tags:
      - Category
    post:
      consumes:
      - multipart/form-data
      description: Create a new product category by specifying its name and optionally
        uploading an image
      parameters:
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      - description: Upload category image (optional)
        in: formData
        name: file
        type: file
      - description: Name of the category
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Category successfully created
          schema:
            $ref: '#/definitions/products.Category'
        "400":
          description: Invalid input or bad request
          schema:
            $ref: '#/definitions/entity.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entity.Error'
      security:
      - ApiKeyAuth: []
      summary: Create Product Category
      tags:
      - Category
  /products/category/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product category by ID
      parameters:
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category successfully deleted
          schema:
            $ref: '#/definitions/products.Message'
        "400":
          description: Invalid input or bad request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete Product Category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Retrieve a product category by ID
      parameters:
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category details
          schema:
            $ref: '#/definitions/products.Category'
        "400":
          description: Invalid input or bad request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Product Category
      tags:
      - Category
    put:
      consumes:
      - multipart/form-data
      description: Update a product category by ID
      parameters:
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Upload category image (optional)
        in: formData
        name: file
        type: file
      - description: Name of the category
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category successfully updated
          schema:
            $ref: '#/definitions/products.Category'
        "400":
          description: Invalid input or bad request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Update Product Category
      tags:
      - Category
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"